strip empty, auto-built nodes
